pyproject.toml
[tool.poetry]
name = "whatsappreminderbot"
version = "0.1.0"
description = "WhatsAppReminderBot with use Redis/celery/Fastapi"
authors = ["Andrey Kazannikov  <andrey.kazannikov@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
redis = "^5.0.6"
celery = "^5.4.0"
flake8 = "^7.0.0"
twilio = "^9.1.1"
flower = "^2.0.1"
pytz = "^2024.1"
pydantic-settings = "^2.3.3"
pydantic = "^2.7.4"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
